requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.28.1
aiosqlite==0.20.0
asyncpg>=0.30.0
aiomysql>=0.2.0
cryptography>=41.0.0
tiktoken